do NEVER EVER compare ULONG against ~0Lsvn statsvn stat
git-svn-id: https://svn.aros.org/svn/aros/trunk@27253 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/necessary/mui/classes/nlist/NList_mcc/NList_func4.c
@@ -680,7 +680,7 @@ ULONG NL_CreateImage(Object *obj,struct NLData *data,Object *imgobj,ULONG flags)
   if (!imgobj)
     return(0);
 
-  if ((flags == ~0L) || !get(imgobj,MUIA_Bitmap_Width,&CI_BM_Width))
+  if ((flags == (ULONG)~0) || !get(imgobj,MUIA_Bitmap_Width,&CI_BM_Width))
     return (NL_CreateImage2(obj,data,imgobj,flags));
 
   if (imgobj && data->SETUP)
